Added src and tests folders to clean repo.
