+++ src/PartialModelSelection.hpp
