+++ tests/PartialModelSelectionMain.cpp
